Fixed DTS table load.

+++ eclipse-workspace/bootloader/src/dev_tree_handler.c
@@ -6,7 +6,7 @@ void * devTreeWrite( void )
 {
 	extern uint8_t dev_tree[];
 	extern int dev_tree_sz;
-	uint32_t * tgs = (uint32_t *)0x10000000;
+	uint8_t * tgs = (uint8_t *)0x10000000;
 	int i;
 	for ( i=0; i<dev_tree_sz; i++ )
 		tgs[i] = dev_tree[i];
